Desenvolvimento/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
cryptography